requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
email-validator

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # For PostgreSQL
# or use: pymysql==1.1.0  # For MySQL

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Computer Vision & ML
mediapipe>=0.10.11
opencv-python==4.8.1.78
ultralytics==8.0.196  # YOLOv8
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
pillow==10.1.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-magic==0.4.27
moviepy==1.0.3
requests==2.31.0
pydantic[email]
